Capture the dialog delete event for the shader preferences dialog
Fixes empty dialog when it is redisplayed.
on_xxx_dialog_delete_event_gtkbuilder_callback() is called when the 
window manager close button is clicked. It is up to this function to 
hide (in our case) the dialog - otherwise nothing happens (return 
gboolean TRUE).
I think that I will need to protect other dialogs in a similar way.