Fixup ellipsoids for displaying atomic ADPs.
The routine gsl_linalg_cholesky_decomp returns a full symmetric
Cholesky decomposition. Applying this with glMultMatrixf gives
distorted ellipsoids. This is particularly visible for ellipsoids with
a large anisotropic ratio and not aligned along the Cartesian axes.

An alternative is to make an eigenvalue and eigenvector decomposition
and then apply those with glScalef and glMultMatrixf, respectively.

This is a first attempt to do this. It does not have a fallback when
HAVE_GSL is not defined. Also here scale eigenvectors by the
square-root of their eigenvalue for simplicity in the calling
sequence.